feat(伪装功能): 重构伪装功能模块并添加窗口图标支持
将伪装功能相关代码从主窗口和设置页面拆分为独立模块 CamouflageFlow
新增窗口图标伪装支持，可自定义应用图标
优化设置页面交互，添加图标预览功能
重构测试用例并增加新测试覆盖
修复设置保存时的潜在竞态条件
