+++ WindBoard/Features/Camouflage/UI/CamouflageSettingsPage.xaml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Page
-    x:Class="WindBoard.Settings.Pages.CamouflageSettingsPage"
+    x:Class="WindBoard.Features.Camouflage.UI.CamouflageSettingsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:l10n="using:WindBoard.Localization"
